Desktop/WebGIS_parma/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.11
django-geojson==3.2.0
django-heroku==0.3.1
django-leaflet==0.28.2
docopt==0.6.2
idna==3.3
jsonfield==3.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
netifaces==0.11.0
numpy==1.19.5
pandas==1.1.5
paramiko==2.9.2
Pillow==8.4.0
plotly==5.5.0
psycopg2==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.5.4
Shapely==1.8.0
six==1.16.0
sqlparse==0.4.2
sshtunnel==0.4.0
tenacity==8.0.1
typing-extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
yarg==0.1.9
